requirements.txt
git+https://github.com/huggingface/transformers
tqdm
mup
pandas
torch
scikit-learn
wandb
kaggle
tiktoken

jax
flax
optax
numpy
tensorflow-datasets
tqdm
tensorflow-probability
gin-config
